% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indEQFEout.r
\name{indEQFEout}
\alias{indEQFEout}
\title{Individual FE estimates for construction of dispersion test}
\usage{
indEQFEout(FullTableFull, YYlist, XXlist)
}
\arguments{
\item{FullTableFull}{data}
}
\value{
Parameter estimates
}
\description{
Individual FE estimates for use constructing in dispersion test, gets 
called as part of constructing the test in kmeans_inner
}
\examples{

}
