% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_inner.r
\name{kmeans_inner}
\alias{kmeans_inner}
\title{K-means algorithm inner}
\usage{
kmeans_inner(nn, FullTableFull, ntreat, Vessel, testvars, nontestvars,
  YYlist, XXlist)
}
\arguments{
\item{FullTableFull}{data}
}
\value{
Converged groupings and associated statistical tests
}
\description{
Keeps regrouping until convergence (original) per seed
}
\examples{

}
