% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_wrapper.r
\name{kmeans_wrapper}
\alias{kmeans_wrapper}
\title{Outside wrapper for full K-means algorithm}
\usage{
kmeans_wrapper(FullTableFull, YYlist, XXlist, Vessel, testvars,
  nontestvars, nseeds, startseeds, ntreat = 1, runparallel = FALSE,
  glimit = 3)
}
\arguments{
\item{FullTableFull}{data}
}
\value{
Converged groupings and associated statistical tests
}
\description{
Outside wrapper
}
\examples{

}
