% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeansout.r
\name{kmeansout}
\alias{kmeansout}
\title{K-means algorithm to find latent groupings}
\usage{
kmeansout(FullTableFull, ntreat, nseeds, Vessel, startseeds, testvars,
  nontestvars, runparallel = FALSE, YYlist, XXlist)
}
\arguments{
\item{FullTableFull}{data}
}
\value{
Converged groupings and associated statistical tests
}
\description{
Keeps regrouping until convergence
}
\examples{

}
