% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooledvcovout.r
\name{pooledvcovout}
\alias{pooledvcovout}
\title{Estimate variance-covariance matrix}
\usage{
pooledvcovout(FullTableFull, betaFE, YYlist, XXlist)
}
\arguments{
\item{FullTableFull}{data}
}
\value{
Variance-covariance matrix
}
\description{
Estimate variance-covariance matrix
}
\examples{

}
