% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDSout.r
\name{testDSout}
\alias{testDSout}
\title{Parameter homogeneity test}
\usage{
testDSout(FullTableFull, pooledvcov, testvars, nontestvars, sigmai, betai,
  betaWFE, Vessel, YYlist, XXlist)
}
\arguments{
\item{FullTableFull}{data}
}
\value{
Parameter homogeneity test statistic
}
\description{
Parameter homogeneity test
}
\examples{

}
