% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackgroupout.r
\name{trackgroupout}
\alias{trackgroupout}
\title{Algorithm to move into closest group}
\usage{
trackgroupout(FullTableFull, pooledvcov, testvars, nontestvars, sigmai,
  betai, betaWFE, TrackGroup, YYlist, XXlist)
}
\arguments{
\item{FullTableFull}{data}
}
\value{
New groupings
}
\description{
Algorithm to move into closest group
}
\examples{

}
